requirements.txt
aiogram==3.13.1
aiohttp==3.10.5
python-dotenv==1.0.1
sqlalchemy==2.0.35
apscheduler==3.10.4
aiosqlite==0.20.0
# buff163_unofficial_api - установить из локальной директории:
# pip install -e ./buff163-unofficial-api

